refactor: argument의 객체 -> 컬렌션으로 변경
